consumer_stock/requirements.txt
sqlalchemy
asyncpg
aiokafka
python-dotenv
